tools/requirements.txt
openpyxl>=3.1.2
python-dotenv>=1.0.0
supabase>=2.4.0
pytest>=8.0.0
